Online image change: handle upgrade of the operator (#145)
This adds upgrade logic for VerticaDB created from older versions of the
operator. We changed the selector label for pods in the sts. The selector label
is immutable, so in order to upgrade to the 1.3.0 release, the sts and their
pods need to be destroyed. This is handled automatically by a new reconcile
actor.

This means that when upgrade to the 1.3.0 release, any running Vertica instance
will be stopped then restarted since deleting sts will cause the pods to go
away.

A new github workflow was added so that we can change operator pod upgrades
going forward.