Support for online upgrade (#150)
This adds support for online upgrade of the Vertica server and allow multiple
subclusters to share the same Service Object.

To initiate an upgrade of a Vertica cluster in Kubernetes you simply change the
name of the container image in the CR. Prior to this stage, we had to drive an
offline upgrade because Vertica didn't support mixed versions running at the
same time. Starting in Vertica 11.1.0 (GA in Feb 2022), a mix of versions is
now supported. Vertica supports this by forcing nodes running the older
versions to be in read-only mode – only queries can be run on the node, no DML
or DDL. We are taking advantage of this feature to implement an online version
of an upgrade.

New externals were added to the CRD to support this mode:

- spec.upgradePolicy. Defines how upgrades will be managed. Available values
  are: Offline, Online and Auto.
  - Offline: means we take down the entire cluster then bring it back up with
    the new image. This is the method of upgrade that the operator supported
    before.
  - Online: will keep the cluster up when the upgrade occurs. The data will go
    into read-only mode until the Vertica nodes from the primary subcluster
    reform the cluster with the new image.
  - Auto: will pick between Offline or Online. Online is only chosen if a
    license Secret exists, the k-Safety of the database is 1 and we are running
    with a Vertica version that supports read-only subclusters.
- spec.temporarySubclusterRouting.names: When doing an online upgrade, we
  designate a subcluster to accept traffic while the other subclusters restart.
  This option is used when you want to reuse an existing subcluster to accept
  traffic. A list of subcluster names is given. The operator will pick the
  first subcluster that is online. You typically would want the first
  subcluster to be a secondary and the second subcluster to some other
  subcluster that will accept traffic when restarting the first.
- spec.temporarySubclusterRouting.template: This specifies that a new
  subcluster be created to accept traffic while a subcluster is down. The
  subcluster will be created at the start of an online upgrade and be removed
  at the end.

New status fields are provided that will allow the upgrade to be monitored. A
summary of the new fields are as follows:

- OnlineUpgradeInProgress: a new status condition was added that will be
  activated when doing an online upgrade. It will be toggled in tandem with the
  existing ImageChangeInProgress status condition. The main purpose of this is
  an indication that the upgrade is being done online.
- OfflineUpgradeInProgress: similar to OnlineUpgradeInProgress except that it
  is set when doing an offline upgrade.
- UpgradeStatus: a human readable message that is updated at various points
  during an upgrade. It is meant to give an indication about what phase the
  upgrade is currently working on. This is maintained for both the online and
  offline upgrades.

The Vertica server needs to be on version 11.1 for this to work. This includes
the current version that we are running and the new version we are upgrading.
So this means online upgrade won't be usable until after the next release. For
instance, you need to be on 11.1.0 already and want to upgrade to 11.1.1.

--- 

Prior to this PR, we had a one-to-one mapping of subclusters to Service
objects. This meant that a single Service object could only direct traffic to a
single subcluster. We needed the ability to have one Service object to direct
to multiple subclusters as part of the work for online upgrade. We are showing
this new feature in a separate section to call this out as it can have uses
that are not specific to online upgrade.

When defining a subcluster in the CR, a new field was added called serviceName.
When the operator reconciles the subclusters in the CR, it will create a
Service object using the name specifed in serviceName. If you want to have
multiple subclusters sharing the same Service object, use the same name for
both. The default behaviour is for each subcluster to have its own Service
object. We do this by using the name of the subcluster as the serviceName.

Here is a sample CR, where multiple subclusters share the same service object:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: sample
spec:
  communal:
    path: "s3://nimbusdb/db"
    endpoint: "http://minio"
    credentialSecret: s3-auth
  subclusters:
    - name: sc1
      size: 1
      serviceName: connections
    - name: sc2
      size: 1
      serviceName: connections

Once created, the following service object will exist that will route
connections between the two subclusters: sample-connections.