Allow env vars to be set in the CR (#146)
This PR set the environment variables in the container at runtime. This is done
by encoding the annotations in the CR as env vars.  For each annotation, all
special characters (not a number a letter or underscore) is turned into
underscore and all letters are turned to uppercase.

For e2e tests update-strategy has been modified and instead of creating new
files, step 60 has been modified to take into account annotations too.