Fix bug that prevented subcluster names with hyphens (#151)
This is a bug fix for when the subcluster name in the CR includes hyphens.
There are a couple of cases where we need to update the default subcluster in
Vertica. The SQL we generated for these commands didn't have quotes around the
name, so they were failing. Failure to change the default cluster prevented
removal subcluster in some cases and messed up the assignment of nodes to
subclusters when creating the database.

Updated the multi-sc testcase (now renamed to multi-sc-sanity) to expose the
issue that the bug fix resolves.