Order the env vars converted from annotations (#164)
This fixes a bug when spec.annotations has multiple annotations. Each
annotation is converted to an environment variable in the server container.
Prior to this fix, if you had multiple annotations defined the StatefulSet we
generate each reconcile cycle could change. This causes the statefulset
controller to go into a rolling upgrade -- update pod, wait for it to be ready,
update next pod, etc.

The StatefulSet wasn't really changing though. It was just the order that we
declared the environment variables. The fix is to sort the environment
variables so that the order is consistent. That way the StatefulSet we generate
doesn't change if the annotations didn't change.