Add back the scripted wait for the webhook
When we upgrade the operator-sdk to 1.17.0, we thought we were getting a
fix for the webhook ready probe.  So we removed the scripted wait and
instead relied on the --wait option through helm.  However, we still saw
cases where the helm command returns after the wait but the webhook
still wasn't 100% ready.  We are adding back the scripted wait to
stabilize things until we can investigate this more.