Add UpgradeRequeueTime logic to prevent exponential backoff during upgrade (#168)
During an online upgrade, we drain all of the connections to a subcluster
before shutting it down. The way that this is instrumented in the code is that
we requeue the current reconcile. This causes the operator to queue up the next
reconcile action using an exponential backoff algorithm. This causes a bad user
experience because the exponential backoff can get up to 16 minutes. This adds
a new spec in our API - upgradeRequeueTime, that helps us configure delaying
the Requeue reconcile.