Allow existing serviceaccount to be used (#171)
Until now, we could only use the serviceaccount (role, rolebinding) provided by
the helm charts. This PR offers the possibility to use a custom serviceaccount
to deploy the operator through helm.

- A new helm parameter (serviceAccountNameOverride) has been added and when it
  is not empty, no new serviceaccount, roles, rolebindings will be created but
  the serviceaccount referenced will be used instead.
- A new helm parameter (webhook.enable) has been added which can disable/enable
  the access to the webhook. This can be useful to allow users with limited
  permissions to install the operator.

Previously, the serviceaccount (the same as the operator) passed to the vertica
pods was constant. Now we added a serviceAccount field to VerticaDBReconciler
that will get the value from a flag and and will use that value in builder.go
to set the correct serviceaccount name.

A new e2e test has been added to check this use case as well as a few samples
to create a custom serviceaccount, role rolebinding.