Upgrade stern version
Upgrading stern to 1.21.0.  I hope this will address the issue we have
with that tool where it can stop logging output partway through an e2e
run.  Stern was added a krew plugin, so I am installing it that way and
removing the tar download we had before.

I am also removing some old dead code with the soak testing.  We never
ran this consistently, and its probably in need of some maintenance to
get working.  So I opted to remove it to clean up the repo.