Fix for a wrong upgradeStatus message
I have noticed that on occasion the upgradeStatus message gets *stuck*
at a stage in the online upgrade.  The upgrade has gone past a stage,
but it will end up reporting the wrong upgradeStatus message.  One
test (online-upgrade-with-transient) is sensitive to this as it checks
the status message in its assertions.  I suspect this occurs because we
are using a stale copy of the VerticaDB when we update the message.  I
am fixing this by making that update more robust.