Fix wrong upgradeStatus message (#184)
I have noticed that on occasion the upgradeStatus message gets *stuck*
at a stage in the online upgrade.  The upgrade has gone past a stage,
but it will end up reporting the wrong upgradeStatus message.  One
test (online-upgrade-with-transient) is sensitive to this as it checks
the status message in its assertions.  I suspect this occurs because we
are using a stale copy of the VerticaDB when we update the message.  I
am fixing this by making that update more robust.

The artifacts are only valuable if the e2e test fails.  I saw a case
where my e2e run was clean but it failed trying to upload the artifacts.
That should be a clean run in my opinion.