Fix timing in online-upgrade-drain (#185)
The e2e testcase online-upgrade-drain can fail intermittently.  This
commit will fix that issue.  The issue it was hitting was that it would
try to connect to vertica through a service object for subcluster pri2.
Due to timing this subcluster might not be totally ready to accept
connections.  Adding a strict assertion that will only continue once
pods in that service object are ready.