Drive object creation ahead of restart reconciler
This fixes a timing hole when deleting StatefulSet.  This was hit
intermittently in the restart-node-multi-sc testcase.  A new testcase
was added to make it more likely to exercise this issue.

The restart reconciler runs before the object reconciler.  The object
reconciler will create the k8s objects but the restart reconciler
depends on those objects.  The reason why we had this order was because
of scale down.  The object reconciler will update the size of any
StatefulSet that are scaling down.  But we need to drive admintools
commands before we do that, which requires Vertica to be up.

We are going to inject a new call to the object reconciler ahead of the
restart reconciler to cover this case.  So that it doesn't do scale down
events, this new run of the object reconciler will only create missing
objects and won't update any.