Handling pending pods better
Two changes to the restart reconciler for pending pods:
- when we do re-ip, lets ignore pods that are pending.  The old
behaviour would requeue them.  They don't have an IP, so there is
nothing we can do
- don't requeue restart if pods are running.  We ended up getting into
an infinite loop without anyway to get out.  Lets continue on with the
reconciler.  I was trying to remove the pending pods but I couldn't
because it was blocked in the restart.