Add a test for operator upgrade from 1.3.1 (#194)
This modifies the existing operator upgrade test from 1.2.0 and uses it as a
template for additional tests. A new test was added to upgrade the operator
from 1.3.1 that uses this new template. The template works as a base for the
test, then we mix in the relevant portions that are specific to that version of
the operator.