Expose prometheus service for operator (#198)
This exposes a /metrics endpoint for prometheus. Prometheus metrics are already
built into the operator-sdk framework that we use. It provides some Go specific
metrics (# of go routines, memory usage, etc.) and some controller specific
metrics (how long the reconciliation iterations are, latency of webhook, etc.).
The work here is exposing the metrics for prometheus.

There are new helm parameters to control the deployment. By default, we expose
the endpoint with an auth proxy sidecar (prometheus.expose =
EnableWithAuthProxy). This sidecar will authenticate each request and only
return the metrics if the caller is using a service account that is linked with
the correct rbac rules.

You can also expose the endpoint without any auth checking (prometheus.expose =
EnableWithoutAuth). This is suitable if you want to make the endpoints
available outside of kubernetes.

There is an option to create a ServiceMonitor CR
(prometheus.createServiceMonitor). This is a CRD that is provided by the
prometheus operator. We don't make this the default because it would add a
dependency on the prometheus operator.

There are a bunch of rbac rules that you need to apply if you want to access
the protected endpoint (prometheus.createProxyRBAC). These have to be cluster
scoped rules, so by default the helm install won't include them. Otherwise, you
cannot install the operator more than once on the same system.