Improved handling of pending pods (#201)
This fixes handling of pending pods. Prior to this PR, we waited for pending
pods to be scheduled before doing any of the reconcile actions in the operator.
A problem with this approach is if you need to do some scaling operator to get
rid of the pending pods, we cannot assume all pending pods will get scheduled.
Perhaps there isn't enough resources on the k8s cluster to ever run the pods.

To handle pending pods better, we now use the Status in the vdb to determine if
a pod has an installation or a database. Previously, we depended on a running
pod. This is key because we can determine if a pending pod should be one we
wait on. For example, if we need to restart we will only wait on the pods that
had previously been part of the cluster. We don't need to wait for any pending
pods that had never run before.

One additional change in here that caused a bunch of files to be touched. When
a k8s event occurs, log an entry to the operator log. This will help with
problem determination.