create db needs to choose a primary subcluster (#202)
When creating a database, we pick a subcluster from the vdb that new nodes will
be added to. We previously picked the very first subcluster, but if this is
defined as a secondary, we end up creating it as a primary. This PR fixes that
by selecting the first primary subcluster in the vdb.

This also will log an event when we detect the shard/node ratio is not optimal.
It should be no higher than 3:1. This could be an important event now that we
have integration with the autoscaler. 

A few small e2e fixes:
- Small timing window exists that cause us to check the service object
  routing before it was actually changed.
- timing window in operator-metrics.  The guage for pod readyiness might
  not yet be updated when we check, even though the pod is ready.
  Adding a retry method.
- Saw failures in upgrade-vertica-ks-0.  If we don't include all of the
  pods in the start_db command, we can get stuck.  Then attempts after
  to get podfacts die because the vsql commands fail.