Do e2e tests when we push to vnext
I see this as being a common pattern, main is frozen while QA tests, but
to continue developing the operator we do new changes on vnext.  This
will become main after the next release.  To aid in testing of this
branch, all pushes to vnext will do the e2e tests now.