Minor e2e fixes
This fixes a bug found in the recent change for pending pod handling.
We no longer waited for all pods to be running before we initiated the
cluster restart.  When k-safety is 0, we need all nodes included when we
restart.  Leaving some out can cause the restart to hang.  This happens
because when k-safety is 0, the absence of one node will cause us to
lose cluster quorum.