Status reconciler to preserve some state (#203)
If pods are pending, we used to lose status state about whether an
installation exists or a db was added at that pod. This would show up
when you did a 'kubectl get vdb' -- the INSTALLED/DBADDED columns could
decrease depending if all pods were running or not.

But we now can infer that state even when the pods aren't running. This
PR will preserve that state so that the status reconciler can report
more accurate information.