Push down more state into /etc/podinfo (#206)
This adds more information to /etc/podinfo. The motivation is that we want the
vertica server to include this stuff in a new k8s specific DC table.

A summary of the new information in /etc/podinfo is as follows:

- k8s-build-date: Build date of the k8s server
- k8s-git-commit: The git commit of the k8s server
- k8s-version: The version of the k8s server
- annotations: A list of all of the annotations set for the pod
- namespace: The namespace the pod is created in
- operator-deployment-method: What method did we use to deploy the operator
- operator-version: What version of the operator are we using

As well, we are exposing the k8s node name and IP as environment variables:

- HOST_IP
- HOST_NODENAME

Two of the states (operator-deployment-method and operator-version) are fixed
for the life of the operator. For this reason they come from a configMap that
is populated either in the helm chart or the OLM bundle.

Also, changed PodFacts so that we have access to a Log.