Allow e2e to be run against private container registry (#208)
This makes some changes to the e2e tests so that we can run it against a
private container registry. Credentials to access the registry are to be
provided in the kustomize config file (names of the variables to set are in
tests/kustomize-defaults.cfg). The credentials are then put in a secret by the
test harness using kustomize.