Fix pod restart with k-safety 0 (#211)
This fixes a regression that was introduced when changing pending pod handling
in this release. When running with k-safety 0, we need to restart all nodes
together. This closes a timing hole where it failed to wait for all of them.
This caused one of the e2e test to fail sporadically.