Enforce required fields in VerticaAutoscaler (#213)
This adds rules to VerticaAutoscaler to make serviceName and
verticaDBName as required.

Also, this relaxes some rules in the webhook wrt serviceName in the
template. I removed the defaulter that had previously set the service
name in the template to match the one in the spec. The webhook only
enforces that the service name match between spec and template if the
template is being used.