Allow vertica server to be run from mounts
This is a change to allow the vertica server to be run from a set of
mount points.  This allows us to change how we develop server + k8s
features.  The old approach was to compile the server, create an RPM,
then package a new container with the RPM.  Using directory mounts that
access the server directly the new approach is much quicker as you only
have to compile the server.  The new changes will get automatically
picked up in k8s because we are using directory mounts directly into the
server repo.

The e2e tests were updated so that we can inject volume/volumeMounts.
This allowed us to run the e2e tests against the server in the directory
mounts.

The kind.sh script was updated so a path to the server repo can be
mounted as a hostPath in the docker container running k8s.