Use pod limits when running admintools (#218)
When you exec into a pod, the ulimits you use are for the container
runtime.  This can differ from the actual limits for the pod/container.
This change ensures we use the pod/container limits when running any
admintools commands.  This ensures the limits are the same for all pods.