Allow vertica server to be run from volume mounts (#217)
This is a change to allow the vertica server to be run from a set of volume
mount points. This allows us to change how we develop server + k8s features.
The old approach was to compile the server, create an RPM, then package a new
container with the RPM. Using directory mounts that access the server directly
the new approach is much quicker as you only have to compile the server. The
new changes will get automatically picked up in k8s because we are using
directory mounts directly into the server repo.

The server container had to be updated. The docker entrypoint, which was
previously in /opt/vertica/bin, had to be moved to another directory. If it was
left where it was, we would lose the entrypoint since we need to mount a path
to /opt/vertica/bin.

The e2e tests were updated so that we can inject volume/volumeMounts. This
allowed us to run the e2e tests against the server in the directory mounts.

You cannot run any of the upgrade tests when using volumeMounts. So moved the
two offline upgrade tests to e2e-online-upgrade and renamed that suite to
e2e-server-upgrade.

The kind.sh script was updated so a path to the server repo can be mounted as a
hostPath in the docker container running k8s.