Prefix operator image with docker.io
This corrects a problem when creating the new catalog entry in
operatorhub.io.  It uses podman to pull the image to verify.  And podman
doesn't assume you are pulling from dockerhub by default.   This change
will make the repo explicit.

We only needed to update the bundle generation, but changing the default
in the helm chart too to be consistent.