Prefix operator image with docker.io (#220)
This corrects a problem when creating the new catalog entry in operatorhub.io.
It uses podman to pull the image to verify. And podman doesn't assume you are
pulling from dockerhub by default. This change will make the repo explicit. I
had to update the CSV manually for the 1.5.0 release. This changes the workflow
to prefix with docker.io so the manual change isn't needed for the next
release.

We only needed to update the bundle generation, but changing the default in the
helm chart too to be consistent.