Allow spread encryption to be set in the VerticaDB CR (#224)
The following parm was added to the VerticaDB CR:

spec.enableSpreadEncryption

When this is set to vertica, then the EncryptSpreadComm parameter will be set
for the database right after it is created. Setting encryption in spread
requires a server restart, which the operator handles automatically. The
default value for this parameter is an empty string, which means no encryption
is set.

The events for the VerticaDB will look like the following when enabling spread
encryption. We will show a successful create database, then a complete server
restart.

Events:
  Type    Reason                   Age   From                Message
  ----    ------                   ----  ----                -------
  Normal  CreateDBStart            78s   verticadb-operator  Calling 'admintools -t create_db'
  Normal  CreateDBSucceeded        48s   verticadb-operator  Successfully created database with subcluster 'my-sc'. It t
ook 30.065841637s
  Normal  StopDBStart              47s   verticadb-operator  Calling 'admintools -t stop_db'
  Normal  StopDBSucceeded          43s   verticadb-operator  Successfully stopped the database.  It took 4s
  Normal  ClusterRestartStarted    42s   verticadb-operator  Calling 'admintools -t start_db' to restart the cluster
  Normal  ClusterRestartSucceeded  33s   verticadb-operator  Successfully called 'admintools -t start_db' and it took 9s

This can only be set for new instances of VerticaDB. It is ignored if the
initPolicy is something other than Create. It only accepts an empty string (to
clear encryption) or vertica. The webhook enforces these rules.

The restart was generalized with a new condition variable:
VerticaRestartNeeded. In the future, if we have other cases where Vertica needs
to be restarted we can use this condition variable.