Don't use docker.io repo when deploying through make
The make target deploy will install using the helm chart.  I am
intentionally clearing the repo to use for the operator image.  That way
the exact image we end up using is the contents of the OPERATOR_IMG
environment variable.  Otherwise, when developing on environments
without kind, we would get pull errors when trying to deploy the
operator.