Log warning if v12.0.0 server and cgroups v2 (#227)
There is an issue in the v12.0.0 server that prevents it from staring in k8s if
running with cgroups v2. This will log a warning event if we detect this
environment. We will still continue to try and start vertica, just in case we
are running with a hotfix that addresses the issue.

I'm changing the default server image to 11.1.1 to make it less likely that we
would hit this issue.