Capture scrutinize in the e2e runs (#229)
In an effort to have more debug information for failed e2e runs, we now collect
scrutinize. Normally the test harness (kuttl) will create a new k8s namespace
for each testcase and delete it at the end. We now delete the namespace
manually, and only if the testcase succeeds. This leaves all of the namespaces
for the failed tests around until the end of the test. We can then collect
scrutinize for each vertica cluster still running.