Bump operator version 1.7.0
This increases the operator to 1.6.0.

We also are adding a new testcase to verify the upgrading of the
operator from 1.6.0.

This fixes an issue where the operator wouldn't update the operator
version in k8s objects (app.kubernetes.io/version).

This also fixes an issue that everytime we updated the operator version
it would force all of the pods to be rescheduled.  This was necessary
because we changed the label for the pod in the sts template portion.  I
now include the version label as a reconciler and not through the sts
template.  The net result is that all of the vertica pods will need to
be rescheduled/restarted when moving to 1.7.0.  But after that version
no restart will be needed.