Missed cleanup of communal on a couple of testcases
Some testcases run a pod to cleanup communal at the start. When they go
to cleanup the communal storage at the end nothing happens because the
pod already exists. We need to delete the clean-communal pod before
doing the cleanup the second time.

We are still seeing the occasional HDFS test fail due to diskfull in
GitHub CI, so I'm hoping this helps.