Allow install_vertica when run on VM in AWS (#240)
This is a small fix to allow the vertica-k8s container to be built on a machine
running in the AWS environment. We run the install to setup the directory
structure for the container. It is detecting we are running in AWS so is
checking other things and failing. The --ignore-install-config option should
bypass the AWS check entirely.