Fix timing in update-strategy
This fixes a small timing window in update-strategy where it was
possible to advance too fast past a test step.  We delete a pod in one step
(50), then suppose to wait for it to come back up in the next step (55).
But we didn't properly wait for the pod to be deleted in step 50.  So,
it was possibly to quickly advance through step 55 before the pod really
came back.