Automatic resize of the PV (#244)
This adds automatic resize of the PV. Whenever the spec.local.requestSize
increases in the VerticaDB, this will trigger increasing the size of each pod's
PV and a resize of the depot inside Vertica. An event will be written for the
VerticaDB when the resize is done.

A few caveats with this:

- The depot must be sized to a percentage of the disk space for it to be
  updated. If it is sized to be a fixed size, the PVC will change but the depot
  won't.
- The storage provisioner must allow volume expansion. Most of the ones used by
  cloud vendors support this. If you are using one such as
  rancher.io/local-path, the default when running k8s with kind, then the PV
  will not increase in size.

Webhook rules were removed to allow the spec.local.requestSize to change. In
theory, only increases to the requestSize should be allowed. But an edge case
exists when changing the size for a storage provisioner that doesn't allow
volume expansion. In such a case, we would want to be able to decrease the
requestSize back to its original value.

A new e2e test was added for this. When run in our CI, which uses
rancher.io/local-path as the default, the test expects the PV not to expand.
When run in an environment that allows volume expansion, then the test uses a
different expected result.

Closes #231