Add feature gate to try out experimental http server in Vertica (#246)
This adds new parameters to the CR to act as a feature gate for the http server
that we are developing in the Vertica server. Its off by default and can only
be enabled if running with the most recent version of Vertica (12.0.1 as of
this PR). The new parameter names are:

- spec.enableHTTPServer: Set this to true to enable the http server. It is
  false by default.
- spec.httpServerSecret: This is a name of a secret that has the TLS
  credentials for the http server. If this is empty, and the http server is
  enabled the operator will generate a secret internally using a self-signed
  cert.

The http server doesn't have many useful functions at the moment. So, it won't
be interesting to any non-Vertica users. For this reason, we won't be
documenting the new parameters.

A new e2e test suite was added. But we intentionally left it out of our CI
because the Vertica version you need (12.0.1) is not yet generally available.