Add support for Java UDx (#248)
The Vertica server image is modified to install jre-headless (version 8) through
the package manager.  This is only done for the full image.  This allows us to
run Java UDx in that version of the image.  The minimal image will continue to
not support Java UDx.  This was done to keep the minimal image small as adding
jre-headless adds about 140mb to the size of the image.

New java udx tests were added but they haven't been added to any CI workflow
yet.  That is to follow.

This also changes the Kubernetes version used in the CI to 1.23.0. This change
was done locally so that tests could be run in a WSL2 environment.