e2e udx tests for python and C++ (#249)
This adds python and C++ tests to the e2e-udx test suite.

Still to be done is to hook up the e2e-udx tests to the GitHub CI.

The cpp UDx test for SourceFunctions.sql is not included for now because errors
occur while vertica is loading those corresponding .so libraries. This might be
an issue resulted from the fact that the lib binaries now are not built within
the context of containers in k8s but rather built elsewhere and uploaded to the
pods.