Don't interrupt a slow Vertica startup
When the operator needs to restart a node, it would kill any running
Vertica process before it attemps to restart.  This was done initially
in case a vertica process that was up, stops resonding.  However, this
behaviour has been problematic if the Vertica startup is just too slow.

This commit will change the behaviour so that it won't uncondionally
kill the Vertica process.  If the startup.log shows that Vertica is in
the process of starting up, then the process is left as is and we
requeue the iteration.