Allow env variables to flow down to Vertica process
This fixes a long standing issue of not being able to set environment
variables that the vertica process will be able to read. The issue was
with the ssh setup. We were using the default config, which prevented
the client from flowing any environment variables. We are opening this
up so the client can flow anything.

We can now set an annotation in the VerticaDB, and it will be set when
the vertica process starts. For instance, this CR:

apiVersion: vertica.com/v1beta1
kind: VerticaDB
metadata:
  name: v
spec:
  annotations:
    TZ: "America/Halifax"
...

Will have the TZ env var set when the vertica process starts.

We will use this feature for other things too, like AWS EKS injecting
environment variables for IRSA.