Some pods may fail to run for a server upgrade change (#271)
This fixes a problem where a server upgrade, accompanied with other CR changes,
may prevent pods from starting. If the image change in the CR along with some
other change that requires a change in the pod spec then the operator was
failing to get the pods running. The change in the pod spec had to be something
that was required for it to run. The example that we hit with the bug is if one
of the mounted volume names change. Such a change requires the pod spec to be
updated to point to the new.

To solve this, we run the ObjReconciler, which is responsible for sts/pod spec
updates, ahead of the upgrade reconcilers.

A new e2e test was added that was able to mimic the error.